Add files via upload
BUG MASTER SYSTEM:
standardize the recording of bugs in systems.
Tables: 

	Bug Master System is designed to standardize the recording of bugs in systems. 
	The system has the following Tables. 
a)	Incident Master: This table will be updated by several roles as roles will update relative table trigger will update the incident master, but only specific views will be available to different roles. The incident master has a backup table using ‘SAVEPOINT’.
b)	Bug Master: The bug master table is designed for the engineer when a bug occurs in the system. This Role will update the Bug ID, System, and ETA in the Bug master table.
c)	Front_Desk: This table is designed for the front_desk who will be handling the customers from the front desk. They will refer the incident ID and enter the customer and status weather the customer is egarly waiting for a solution or agreed to grant some time for resolution this particular input will update the priority of the bug.  
d)	Quality Team: Quality Team will update the resolved time total downtime and engineer who worked on  the incident.
e)	Engineering Master table will be updated by engineer for related BUG ID and system and short resolution steps or steps of troubleshooting and resolving the bug.
f)	Engieering SME will  refer the incident ID and BugID to provide the steps required to remove the BUG from system. 
 


Views: Scematic views for specific user/role:

	Front_Desk_VIEW : Incident_ID, Active Bug, related customer, ETA and standard ETA columns are taken from the incident master table as a view. ETA is expected time of resolution and Standard ETA is average downtime of the bug from history calculated by aggregate functions and group by. 
	Engineer_View: Engineer can see BugID and  engineer assigned for resolving the particular bug. This role has access  for the Knowledge_Base table of the Bug_ID and Resolution.
	Quality Team can see Engineer assigned for active bug and ETA , steps required to resolve particular BUG.